proxy aware factory
